Moved content of TwitterCreationService to TwitterResourceService
also moved TwitterKeys to configuration package
